Few modifications to not get errors
In Debian 8.0 with Perl 5.20.2-3 and expect 5.45-6 there where errors without
my modifications.

The errors for the csbrancid because of perl is:

Trying to get all of the configs.
defined(%hash) is deprecated at /usr/lib/rancid/bin/csbrancid line 72.
        (Maybe you should just omit the defined()?)

The error for the expect was:

Trying to get all of the configs.
/usr/bin/env: expect --: Does not exist the file or directory